Add module init; move symbols from macros + export
