+++ gdnative/src/lib.rs
@@ -58,14 +58,11 @@
 // Items, which are #[doc(hidden)] in their original crate and re-exported with a wildcard, lose
 // their hidden status. Re-exporting them manually and hiding the wildcard solves this.
 #[doc(inline)]
-pub use gdnative_core::{core_types, export, log, object, InitializeInfo, TerminateInfo};
+pub use gdnative_core::{core_types, export, init, log, object};
 
 /// Collection of declarative `godot_*` macros, mostly for GDNative registration and output.
 pub mod macros {
-    pub use gdnative_core::{
-        godot_dbg, godot_error, godot_gdnative_init, godot_gdnative_terminate, godot_init,
-        godot_nativescript_init, godot_print, godot_warn, godot_wrap_method,
-    };
+    pub use gdnative_core::{godot_dbg, godot_error, godot_print, godot_warn};
 }
 
 // Implementation details (e.g. used by macros).
@@ -85,7 +82,7 @@ pub mod prelude;
 #[cfg(feature = "bindings")]
 pub use gdnative_bindings as api;
 
+/// Support for async code
 #[doc(inline)]
 #[cfg(feature = "async")]
-/// Support for async code
 pub use gdnative_async as tasks;
